Add an Order model to represent a customer order of products
The model handles the details of an order, including the list of
order items, the associated charge (if any), the customer making the
order, etc. Include getters, setters, equals() and hashCode() methods,
a JSON string constructor, and JSON string serializer methods for
fields of type ObjectNode which will be needed for object mapping.