Update the Quota model to include usage
To streamline accessing quotas and usage on a per-Subject basis, add
a usage attribute to the quota model.  This field will get updated
periodically in the database so a call to getQuotas() will include
usage for the Subject. Update the database schema and SQL DDL scripts,
along with tests, helpers, and mappers that are affected by the change.